Check sigint when a worker catches an error
Check the sigint flag when a fuzz worker detects a crash. This fixes a
bug where killing the jsfuzz during the execution of the fuzz function
can cause an error as well. This bug is particularly likely to happen if
you're fuzzing a slow & blocking function.

With this change, if the sigint flag is set, the worker will close
gracefully. If the sigint flag is not set, the worker will behave the
same as before.

The `setTimeout` is necessary as the sigint flag is not guaranteed to
have been switched before the worker reaches the except block. I set it
to 10 as neither 0 or 1 seemed to be sufficient (i.e. I was still able
to cause a crash manually).