Import yargs only once
In index.ts, import yargs (the CLI arguments parser) once at the top of
the module and use it like that. For this to work, the import statement
had to be changed to a wildcard import because of a TypeScript compile
error with `yargs.scriptName("jsfuzz")` when using the default import.